Safari-611.1.10/LayoutTests/webanimations/keyframe-effect-target-kept-alive-expected.txt

This test checks that a keyframe effect keeps its target alive even if that node has no other reference.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS effect.target.localName is "div"
PASS successfullyParsed is true

TEST COMPLETE
